Allow for analysis of a single user at a time
Rather than analyze the followers and following for a specific user, analyze the
user themselves to see the gender of that specific user as determined by the
algorithm.

This is useful in helping to make the algorithm better and to give people the
ability to see how they are being gendered by this project.