Alter priority order of textual gender markers
The first heuristic for the algorithm is based on the text of the account's bio. This change alters the priority ordering of the words used for that heuristic. In particular, the old priority order had pronouns (e.g. "they") before other descriptors (e.g. "nonbinary"). This commit reverses that. It also puts the more direct descriptors, like "man," before the less direct descriptors, like "father."